requirements.txt
requests
websocket-client
cryptography